Implement Skip List data structure with memory arena integration and add related tests

+++ internal/storage/sepia/internal/marena/marena_test.go
@@ -27,7 +27,7 @@ func TestAllocate(t *testing.T) {
 		t.Fatal("Allocation returned invalid address")
 	}
 	gotSize := marena.Size(addr)
-	wantSize := 100
+	wantSize := uint32(100)
 	if gotSize != wantSize {
 		t.Errorf("Size(addr) = %d, want %d", gotSize, wantSize)
 	}
